% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allMethods}
\alias{allMethods}
\title{Machine learning model table for mlComb()
Includes machine learning models used for the mlComb function}
\format{
A data frame with 113 rows and 2 variables:
\describe{
\item{Method}{Valid name for the function}
\item{Model}{Model name}
}
}
\usage{
data(allMethods)
}
\description{
Machine learning model table for mlComb()
Includes machine learning models used for the mlComb function
}
\examples{
data(allMethods)
allMethods

}
\keyword{datasets}
