% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlComb.R
\name{availableMethods}
\alias{availableMethods}
\title{Available classification/regression methods in \code{dtComb}}
\usage{
availableMethods()
}
\description{
This function returns a data.frame of available classification
methods in \code{dtComb}. These methods are imported from caret package.
}
\examples{

availableMethods()

}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec
}
