% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData2}
\alias{exampleData2}
\title{Biomarker data from carriers of a rare genetic disorder
A data set containing the carriers of a rare genetic disorder for 120 samples.}
\format{
A data frame with 120 rows and 5 variables:
\describe{
\item{Group}{Indicator whether the person is a carrier; values:
\code{"carriers"} and \code{"normals"}.}
\item{m1}{Biomarker 1, 1. measurement blood sample}
\item{m2}{Biomarker 2, 2. measurement blood sample}
\item{m3}{Biomarker 3, 3. measurement blood sample}
\item{m4}{Biomarker 4, 4. measurement blood sample}
}
}
\usage{
data(exampleData2)
}
\description{
Biomarker data from carriers of a rare genetic disorder
A data set containing the carriers of a rare genetic disorder for 120 samples.
}
\examples{
data(exampleData2)
exampleData2$Group <- factor(exampleData2$Group)
gcol <- c("#E69F00", "#56B4E9")
plot(exampleData2$m1, exampleData2$m2,
  col = gcol[as.numeric(exampleData2$Group)]
)

}
\keyword{datasets}
