% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData3}
\alias{exampleData3}
\title{A simulation data containing 250 diseased and 250 healthy individuals.}
\format{
A data frame with 500 rows and 3 variables:
\describe{
\item{status}{Indicator of one's condition, values healthy and diseased}
\item{marker1}{1. biomarker}
\item{marker2}{2. biomarker}
}
}
\usage{
data(exampleData3)
}
\description{
A simulation data containing 250 diseased and 250 healthy individuals.
}
\examples{
data(exampleData3)
exampleData3$status <- factor(exampleData3$status)
gcol <- c("#E69F00", "#56B4E9")
plot(exampleData3$marker1, exampleData3$marker2,
  col = gcol[as.numeric(exampleData3$status)]
)

}
\keyword{datasets}
