% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linComb.R
\name{helper_TS}
\alias{helper_TS}
\title{AUC calculator for TS method.}
\usage{
helper_TS(theta, markers, status)
}
\arguments{
\item{theta}{a \code{numeric} parameter that will be estimated in the TS
method for the combination score}

\item{status}{a \code{factor} data frame that includes the actual disease
status of the patients}

\item{marker.set}{a \code{numeric} data frame that contains the biomarkers}
}
\value{
A \code{numeric} AUC value calculated with combination scores using
trigonometric functions of theta
}
\description{
The \code{helper_TS} function calculates the combination
coefficient and AUC value of given biomarkers for the TS method
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- cbind(exampleData1$ddimer, exampleData1$log_leukocyte)
status <- factor(exampleData1$group, levels = c("not_needed", "needed"))

theta <- 0.08

stat <- helper_TS(theta = theta, markers = markers, status = status)

}
\author{
Serra Bersan Gengec, Ilayda Serra Yerlitas
}
