% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathComb.R
\name{mathComb}
\alias{mathComb}
\title{Combine two diagnostic tests with several mathematical operators and
distance measures.}
\usage{
mathComb(
  markers = NULL,
  status = NULL,
  event = NULL,
  method = c("add", "multiply", "divide", "subtract", "distance", "baseinexp",
    "expinbase"),
  distance = c("euclidean", "manhattan", "chebyshev", "kulczynski_d", "lorentzian",
    "avg", "taneja", "kumar-johnson"),
  standardize = c("none", "range", "zScore", "tScore", "mean", "deviance"),
  transform = c("none", "log", "exp", "sin", "cos"),
  power.transform = FALSE,
  direction = c("auto", "<", ">"),
  conf.level = 0.95,
  cutoff.method = c("youden", "roc01")
)
}
\arguments{
\item{markers}{a \code{numeric} data frame that includes two diagnostic tests
results}

\item{status}{a \code{factor} vector that includes the actual disease
status of the patients}

\item{event}{a \code{character} string that indicates the event in the status
to be considered as positive event}

\item{distance}{a \code{character} string specifying the method used for
combining the markers. The available methods are:
\itemize{
\item \code{euclidean}: The euclidean distance is (named after Euclid) a
straight line distance between two points. Euclid argued that the shortest
distance between two points is always a line.
Euclidean: d = sqrt( ∑ | P_i - Q_i |^2)
\item \code{manhattan}: The Manhattan distance, also called the Taxicab
distance or the City Block distance, calculates the distance between two
real-valued vectors.
Manhattan : d = ∑ | P_i - Q_i |
\item \code{chebyshev}: Chebyshev distance (or Tchebychev distance), maximum
metric, or L∞ metric is a metric defined on a vector space where the
distance between two vectors is the greatest of their differences along any
coordinate dimension.
Chebyshev : $d = max | P_i - Q_i |$
\item \code{kulczynski_d}:
Kulczynski d : d = ∑ | P_i - Q_i | / ∑ min(P_i , Q_i)
\item \code{lorentzian}:
Lorentzian : d = ∑ ln(1 + | P_i - Q_i |)
\item \code{avg}:
Avg(L_1, L_n) : d = ∑ | P_i - Q_i| + max{ | P_i - Q_i |} / 2
\item \code{taneja}:
Taneja : d = ∑ ( P_i + Q_i / 2) * log( P_i + Q_i / ( 2 * sqrt( P_i * Q_i)) )
\item \code{kumar-johnson}:
Kumar-Johnson : d = ∑ (P_i^2 - Q_i^2)^2 / 2 * (P_i * Q_i)^1.5
}}

\item{standardize}{a \code{character} string indicating the name of the
standardization method. The default option is no standardization applied.
Available options are:
\itemize{
\item \code{range}: Standardization to a range between 0 and 1
\item \code{zScore}: Standardization using z scores with mean = 0
and standard deviation = 1
\item \code{tScore}: Standardization using T scores. The range varies
between usually 20 and 80
\item \code{mean}: Standardization with sample mean = 1
\item \code{deviance}: Standardization with sample standard deviation = 1
}}

\item{transform}{a \code{character} string indicating the name of the
standardization method. The default option is no standardization applied.
Available options are:
\itemize{
\item \code{log}: Applies logarithm transform to markers before calculating
combination score
\item \code{exp}: Applies exponential transform to markers before
calculating combination score
\item \code{sin}: Applies sinus trigonometric transform to markers before
calculatin combination score
\item \code{cos}: Applies cosinus trigonometric transform to markers before
calculating combination score
}}

\item{power.transform}{a \code{logical} variable that determines whether to
apply power to the markers giving the optimum AUC value in the \link{-3, 3}
range, before calculating the combination score (FALSE, default).}

\item{direction}{a \code{character} string determines in which direction the
comparison will be made.  “>”: if the predictor values for the control group
are higher than the values of the case group (controls > cases).
“<”: if the predictor values for the control group are lower or equal than
the values of the case group (controls < cases).}

\item{conf.level}{a \code{numeric} values determines the confidens interval
for the roc curve(0.95, default).}

\item{cutoff.method}{a \code{character} string determines the cutoff method
for the roc curve.}
}
\value{
A list of \code{numeric} linear combination scores calculated
according to the given method and standardization option
}
\description{
The code{mathComb} function returns the combination results of
two diagnostic tests with different mathematical operators, distance
measures, standardization, and transform options.
}
\examples{

data(exampleData1)
markers <- exampleData1[, -1]
status <- factor(exampleData1$group, levels = c("not_needed", "needed"))
event <- "needed"
direction <- "<"
cutoff.method <- "youden"

score1 <- mathComb(markers = markers, status = status, event = event,
method = "distance", distance ="euclidean", direction = "auto", standardize = "range",
cutoff.method = cutoff.method)

score2 <- mathComb(markers = markers, status = status, event = event,
method = "expinbase", transform = "cos", direction = direction,
cutoff.method = cutoff.method)

score3 <- mathComb(markers = markers, status = status, event = event,
method = "add", transform = "exp", direction = direction, power.transform = TRUE,
cutoff.method = cutoff.method)

}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec
}
