% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinComb.R
\name{nonlinComb}
\alias{nonlinComb}
\title{Combine two diagnostic tests with several non-linear combination methods.}
\usage{
nonlinComb(
  markers = NULL,
  status = NULL,
  event = NULL,
  method = c("polyreg", "ridgereg", "lassoreg", "elasticreg", "splines", "sgam",
    "nsgam"),
  degree1 = 3,
  degree2 = 3,
  df1 = 4,
  df2 = 4,
  resample = c("none", "cv", "repeatedcv", "boot"),
  nfolds = 5,
  nrepeats = 3,
  niters = 10,
  standardize = c("none", "range", "zScore", "tScore", "mean", "deviance"),
  include.interact = FALSE,
  alpha = 0.5,
  show.plot = TRUE,
  direction = c("auto", "<", ">"),
  conf.level = 0.95,
  cutoff.method = c("youden", "roc01"),
  ...
)
}
\arguments{
\item{markers}{a \code{numeric} data frame that includes two diagnostic tests
results}

\item{status}{a \code{factor} vector that includes the actual disease
status of the patients}

\item{event}{a \code{character} string that indicates the event in the status
to be considered as positive event}

\item{method}{a \code{character} string specifying the method used for
combining the markers. The available methods are:
\itemize{
\item \code{polyreg}: The method builds a logistic regression model with the
feature space created and returns the probability of a positive event for
each observation.
\item \code{ridgereg}: Ridge regression is a shrinkage method used to
estimate the coefficients of highly correlated variables and in this case
the polynomial feature space created from two biomarkers. For the
implementation of the method, glmnet library is used with two functions:
cv.glmnet() to run a cross validation  model to determine the tuning
parameter λ and glmnet() to fit the model with the selected tuning parameter.
\item \code{lassoreg}: Lasso regression is also a shrinkage method with one
difference is that at the end this method returns the coefficients of some
features as 0, makes this method useful for feature elimination as well.
The implementation is similar to Ridge regression, cross validation for
parameter selection and model fit are implemented with glmnet library.
\item \code{elasticreg}: Elastic Net regression is obtained by combining the
penalties of Ridge and Lasso regression to get the best of both models. The
model again includes a tuning parameter λ as well as a mixing parameter α
taken form the user which takes a value between 0 (ridge) and 1 (lasso) to
determine the weights of the loss functions of Ridge and Lasso regressions.
\item \code{splines}: With the applications of regression models in a
polynomial feature space the second non-linear approach to combining
biomarkers comes from applying several regression models to the data set
using a function derived from pecewise polynomials. Splines are implemented
with degrees of freedom and degrees of the fitted polynomials taken from the
user. For the implementation splines library is used to build piecewise
logistic regression models with base splines.
\item \code{sgam}: In addition to the basic spline structure, Generalized
Additive Models are applied with smoothing splines using the gam
library R.
\item \code{nsgam}: In addition to the basic spline structure, Generalized
Additive Models are applied with natural cubic splines using the gam
library R.
}}

\item{degree1}{a \code{numeric} value for polynomial based methods indicates
the degree of the feature space created for marker 1, for spline based
methods the degree of the fitted polynomial between each node for marker 1.
(3, default)}

\item{degree2}{a \code{numeric} value for polynomial based methods indicates
the degree of the feature space created for marker 2, for spline based
methods the degree of the fitted polynomial between each node for marker 2
(3, default)}

\item{df1}{a \code{numeric} value that indicates the number of knots as the
degrees of freedom in spline based methods for marker 1 (4, default)}

\item{df2}{a \code{numeric} value that indicates the number of knots as the
degrees of freedom in spline based methods for marker 2 (4, default)}

\item{resample}{a \code{character} string indicating the name of the
resampling options. Bootstrapping Cross validation and repeated cross
validation are given as the options for resampling along with the number
of folds and number of repeats.
\itemize{
\item \code{boot}: Bootstrapping is performed in a similar way; the dataset
is divided into folds with replacement and models are trained and tested
in these folds to determine the best parameters for the given method and
dataset.
\item \code{cv}: Cross-validation resampling, the dataset is divided into
the number of folds given without replacement, in each iteration one fold
is selected as the test set, the model is built using the remaining folds
and tested on the test set. The corresponding AUC values are kept in a list
as well as the parameters used for the combination. The best performed model
is selected and the combination score returned for the whole dataset.
\item \code{repeatedcv}: repeated cross-validation the process is repeated
and the best performed models selected at each step is stored in another
list, the best performed among these models is selected to be applied to
the entire dataset.
}}

\item{nfolds}{a \code{numeric} value that indicates the number of folds for
cross validation based resampling methods  (5, default)}

\item{nrepeats}{a \code{numeric} value that indicates the number of repeats
for "repeatedcv" option of resampling methods (3, default)}

\item{niters}{a \code{numeric} value that indicates the number of
bootstrapped resampling iterations (10, default)}

\item{standardize}{a \code{character} string indicating the name of the
standardization method. The default option is no standardization applied.
Available options are:
\itemize{
\item \code{range}: Standardization to a range between 0 and 1
\item \code{zScore}: Standardization using z scores with mean = 0
and standard deviation = 1
\item \code{tScore}: Standardization using T scores. The range varies between
usually 20 and 80
\item \code{mean}: Standardization with sample mean = 1
\item \code{deviance}: Standardization with sample standard deviation = 1
}}

\item{include.interact}{a \code{logical} indicator that specifies whether to
include the interaction between the markers to the feature space created for
polynomial based methods (FALSE, default)}

\item{alpha}{a \code{numeric} value as the mixing parameter in Elastic Net
Regression method (0.5, default)}

\item{show.plot}{a \code{logical} a \code{logical}. If TRUE, a ROC curve is
plotted. Default is TRUE}

\item{direction}{a \code{character} string determines in which direction the
comparison will be made.  “>”: if the predictor values for the control group
are higher than the values of the case group (controls > cases).
“<”: if the predictor values for the control group are lower or equal than
the values of the case group (controls < cases).}

\item{conf.level}{a \code{numeric} values determines the confidence interval
for the ROC curve(0.95, default).}

\item{cutoff.method}{a \code{character} string determines the cutoff method
for the ROC curve.}
}
\value{
A list of \code{numeric} nonlinear combination scores calculated
according to the given method and standardization option
}
\description{
The \code{nonlinComb} function calculates the combination
scores of two diagnostic tests selected among several non-linear combination
methods and standardization options
}
\examples{
data("exampleData1")
data <- exampleData1

markers <- data[, -1]
status <- factor(data$group, levels = c("not_needed", "needed"))
event <- "needed"
direction <- "<"
cutoff.method <- "youden"

score1 <- nonlinComb(markers = markers, status = status, event = event,
method = "ridgereg",  include.interact = FALSE, cutoff.method = "youden",  
direction = "<")

score2 <- nonlinComb(markers = markers, status = status, event = event,
method = "splines", resample = "none", cutoff.method = "youden",
standardize = "tScore", direction = "<")

score3 <- nonlinComb(markers = markers, status = status, event = event,
method = "lassoreg", resample = "repeatedcv", include.interact = TRUE,
cutoff.method = "youden", standardize = "zScore", direction = "auto")

}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec
}
