% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathComb.R
\name{power.add}
\alias{power.add}
\title{The power giving optimum AUC value for the addition operator}
\usage{
power.add(n, marker.set)
}
\arguments{
\item{n}{a \code{numeric} parameter to estimate the best power.transform for
the combination score}

\item{marker.set}{a \code{numeric} data frame that contains the biomarkers}
}
\value{
A \code{numeric} combination score value for calculated with the
help of exponent
}
\description{
The \code{power.add} function calculates the nth power of the
markers for n selected in the range -3 <= n <= 3 for the addition function.
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- cbind(exampleData1$ddimer, exampleData1$log_leukocyte)
status <- factor(exampleData1$group, levels = c("not_needed", "needed"))

n = 0.5

comb.score <- power.add(n, marker.set = markers)

}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec
}
