% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predComb.R
\name{predComb}
\alias{predComb}
\title{Predict combination scores and labels for new data sets using the
training model}
\usage{
predComb(model, newdata)
}
\arguments{
\item{model}{a \code{list} object where the parameters from the training
model are saved.}

\item{newdata}{a \code{numeric} new data set that includes biomarkers that
have not been introduced to the model before.}
}
\value{
A \code{data.frame} predicted combination scores (or probabilities)
and labels
}
\description{
The \code{predComb} a function that generates predictions
for a new dataset of biomarkers using the parameters from the fitted model.
The function takes arguments newdata and model. The output of the function is
the combination scores and labels of object type.
}
\examples{

#call data
data(exampleData1)

#define the function parameters
markers <- exampleData1[, -1]
status <- factor(exampleData1$group, levels = c("not_needed", "needed"))
event <- "needed"

score1 <- linComb(markers = markers, status = status, event = event,
method = "logistic", resample= "none",
standardize = "none", direction = "<", cutoff.method = "youden")

comb.score1 <- predComb(score1, markers)

score2 <- nonlinComb(markers = markers, status = status, event = event,
method = "nsgam", resample = "cv", include.interact = FALSE, direction = "<",
standardize = "zScore", cutoff.method = "youden")

comb.score2 <-  predComb(score2, markers)

score3 <- mathComb(markers = markers, status = status, event = event,
method = "distance", distance ="euclidean", direction = "auto", 
standardize = "tScore", cutoff.method = "youden")

comb.score3 <-  predComb(score3, markers)

}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec
}
