% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.train.R
\name{print_train}
\alias{print_train}
\title{Standardization with respect to the t score.}
\usage{
print_train(print_model)
}
\arguments{
\item{data}{a \code{numeric} dataframe of biomarkers}

\item{standardize}{a a \code{string} cdscfvf}
}
\value{
A \code{numeric} dataframe of standardized biomarkers
}
\description{
The \code{std.train} standardizes using T scores. The range
varies between usually 20 and 80
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- exampleData1[, -1]
markers2 <- std.train(markers, tScore)

}
\author{
Serra Ilayda  Yerlitas, Serra Bersan Gengec
}
