% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{std.mean}
\alias{std.mean}
\title{Standardization with respect to the sample mean.}
\usage{
std.mean(markers)
}
\arguments{
\item{markers}{a \code{numeric} data frame of biomarkers}
}
\value{
A \code{numeric} data frame of standardized biomarkers
}
\description{
The \code{std.mean} standardizes with sample mean = 1
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- cbind(exampleData1$ddimer, exampleData1$log_leukocyte)
markers2 <- std.mean(markers)

}
\author{
Serra Bersan Gengec, Ilayda Serra Yerlitas
}
