% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{std.range}
\alias{std.range}
\title{Standardization with respect to range.}
\usage{
std.range(markers)
}
\arguments{
\item{markers}{a \code{numeric} data frame of biomarkers}
}
\value{
A \code{numeric} data frame of standardized biomarkers
}
\description{
The \code{std.range} standardizes to a range between 0 and 1
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- cbind(exampleData1$ddimer, exampleData1$log_leukocyte)
markers2 <- std.range(markers)

}
\author{
Serra Bersan Gengec, Ilayda Serra Yerlitas
}
