% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{std.test}
\alias{std.test}
\title{Standardization with respect to the t score.}
\usage{
std.test(newdata, model)
}
\arguments{
\item{newdata}{a \code{numeric} dataframe of biomarkers}

\item{model}{a \code{string} dcdcdcdz}
}
\value{
A \code{numeric} dataframe of standardized biomarkers
}
\description{
The \code{std.test} standardizes using T scores. The range
varies between usually 20 and 80
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- exampleData1[, -1]
markers2 <- std.test(markers, score1)

}
\author{
Serra Ilayda  Yerlitas, Serra Bersan Gengec
}
