% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{std.tscore}
\alias{std.tscore}
\title{Standardization with respect to the t score.}
\usage{
std.tscore(markers)
}
\arguments{
\item{markers}{a \code{numeric} dataframe of biomarkers}
}
\value{
A \code{numeric} dataframe of standardized biomarkers
}
\description{
The \code{std.tscore} standardizes using T scores. The range
varies between usually 20 and 80
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- cbind(exampleData1$ddimer, exampleData1$log_leukocyte)
markers2 <- std.tscore(markers)

}
\author{
Serra Bersan Gengec, Ilayda Serra Yerlitas
}
