% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{std.zscore}
\alias{std.zscore}
\title{Standardization with respect to z score.}
\usage{
std.zscore(markers)
}
\arguments{
\item{markers}{a \code{numeric} data frame of biomarkers}
}
\value{
A \code{numeric} data frame of standardized biomarkers
}
\description{
The \code{std.zscore} standardizes using z scores with
mean = 0 and standard deviation = 1
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- cbind(exampleData1$ddimer, exampleData1$log_leukocyte)
markers2 <- std.zscore(markers)

}
\author{
Serra Bersan Gengec, Ilayda Serra Yerlitas
}
