% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlComb.R
\name{availableMethods}
\alias{availableMethods}
\title{Available classification/regression methods in \code{dtComb}}
\usage{
availableMethods()
}
\value{
\code{No return value} contains the method names and explanations of the
machine-learning models available for the dtComb package.
}
\description{
This function returns a data.frame of available classification
methods in \code{dtComb}. These methods are imported from the caret package.
}
\examples{

availableMethods()

}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec, Necla Kochan,
Gozde Erturk Zararsiz, Selcuk Korkmaz, Gokmen Zararsiz
}
