% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtComb.R
\docType{package}
\name{dtComb}
\alias{dtComb}
\title{dtComb: A Comprehensive R Library for Combining Diagnostic Tests}
\description{
The dtComb package calculates combination scores of two biomarkers given
under four main categories: linear combinations with the linComb function,
non-linear combinations with the nonlinComb function,
mathematical operators with the mathComb function, and machine learning
algorithms with the mlComb function.
}
