% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtComb.R
\docType{package}
\name{dtComb}
\alias{dtComb-package}
\alias{dtComb}
\title{dtComb: A Comprehensive R Package for Combining Diagnostic Tests}
\description{
The \pkg{dtComb} package provides tools to calculate combination scores of
two biomarkers, categorized into four main types:
\itemize{
\item \code{linComb}: Linear combinations,
\item \code{nonlinComb}: Non-linear combinations,
\item \code{mathComb}: Mathematical operator-based combinations,
\item \code{mlComb}: Machine learning–based combinations.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gokmenzararsiz/dtComb}
}

}
\author{
\strong{Maintainer}: Gokmen Zararsiz \email{gokmen.zararsiz@gmail.com} [contributor]

Authors:
\itemize{
  \item Serra Ilayda Yerlitas \email{ilaydayerlitas340@gmail.com} [contributor]
  \item Serra Bersan Gengec \email{serrabersan@gmail.com} [contributor]
  \item Necla Kochan \email{necla.kayaalp@gmail.com} [contributor]
  \item Gozde Erturk Zararsiz \email{gozdeerturk9@gmail.com} [contributor]
  \item Selcuk Korkmaz \email{selcukorkmaz@gmail.com} [contributor]
}

}
