% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printtrain.R
\name{kappa.accuracy}
\alias{kappa.accuracy}
\title{Calculate Cohen's kappa and accuracy.}
\usage{
\method{kappa}{accuracy}(DiagStatCombined)
}
\arguments{
\item{DiagStatCombined}{a \code{numeric} table of confusion matrix of the
calculated combination score.}
}
\value{
A \code{list} of Cohen's kappa and accuracy values
}
\description{
The \code{kappa.accuracy} calculates Cohen's kappa and accuracy.
}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec, Necla Kochan,
Gozde Erturk Zararsiz, Selcuk Korkmaz, Gokmen Zararsiz
}
