% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linComb.R
\name{linComb}
\alias{linComb}
\title{Combine two diagnostic tests with several linear combination methods.}
\usage{
linComb(
  markers = NULL,
  status = NULL,
  event = NULL,
  method = c("scoring", "SL", "logistic", "minmax", "PT", "PCL", "minimax", "TS"),
  resample = c("none", "cv", "repeatedcv", "boot"),
  nfolds = 5,
  nrepeats = 3,
  niters = 10,
  standardize = c("none", "range", "zScore", "tScore", "mean", "deviance"),
  ndigits = 0,
  direction = c("auto", "<", ">"),
  conf.level = 0.95,
  cutoff.method = c("youden", "roc01")
)
}
\arguments{
\item{markers}{a \code{numeric} data frame that includes two diagnostic tests
results}

\item{status}{a \code{factor} vector that includes the actual disease
status of the patients}

\item{event}{a \code{character} string that indicates the event in the status
to be considered as positive event}

\item{method}{a \code{character} string specifying the method used for
combining the markers. The available methods are:
\itemize{
\item \code{scoring}: Combination score obtained using the slope values of
the relevant logistic regression model
\item \code{SL}: Su and Liu combination score obtained by using Fisher's
discriminant function under the assumption of multivariate normal
distribution model and proportionate covariance matrices
\item \code{logistic}: Combination score obtained by fitting a logistic
regression model
\item \code{minmax}: Linearly combines the minimum and maximum values of
the markers
\item \code{PT}: Pepe and Thompson combination score obtained by
proportioning the slope values the relevant logistic regression model
\item \code{PCT}: Pepe, Cai and Langton combination score obtained by using
AUC as the parameter of a logistic regression model
\item \code{minimax}: Combination score obtained with the Minimax procedure
\item \code{TS}: Combination score obtained by using the trigonometric
functions of the theta value that optimizes the AUC
}
\bold{IMPORTANT}: See Details for further information.}

\item{standardize}{a \code{character} string indicating the name of the
standardization method. The default option is no standardization applied.
Available options are:
\itemize{
\item \code{range}: Standardization to a range between 0 and 1
\item \code{zScore}: Standardization using z scores with mean = 0
and standard deviation = 1
\item \code{tScore}: Standardization using T scores. The range varies between
usually 20 and 80
\item \code{mean}: Standardization with sample mean = 1
\item \code{deviance}: Standardization with sample standard deviation = 1
}}

\item{ndigits}{a \code{integer} value to indicate the number of decimal places
to be used for rounding in Scoring method}

\item{init.param}{a \code{numeric} initial value to be used for optimization
in minmax, PCL, minimax and TS methods}
}
\value{
A list of \code{numeric} linear combination scores calculated
according to the given method and standardization option
}
\description{
The \code{linComb} function calculates the combination
scores of two diagnostic tests selected among several linear combination
methods and standardization options
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- exampleData1[, -1]
status <- factor(exampleData1$group, levels = c("not_needed", "needed"))
event <- "needed"

score1 <- linComb(markers = markers, status = status, event = event,
method = "scoring", resample= "none",
standardize = "none", direction = "<", cutoff.method = "youden")

score2 <- linComb(markers = markers, status = status, event = event,
method = "TS", resample = "boot", standardize = "zScore", direction = "<", 
cutoff.method = "youden")

score3 <- linComb(markers = markers, status = status, event = event,
method = "logistic", resample = "repeatedcv", direction = "<", 
cutoff.method = "youden")

}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec
}
