% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printtrain.R
\name{print_train}
\alias{print_train}
\title{Print the summary of linComb, nonlinComb, mlComb and mathComb
functions.}
\usage{
print_train(print_model)
}
\arguments{
\item{print_model}{a \code{list} of parameters taken from the fitted model
that includes the combination method, resampling method, pre-processing
method, selected optimum parameters and the results of fit.}
}
\value{
\code{No return value} writes a summary of the results to the console.
}
\description{
The \code{print_train} function prints the summary statistics of
the fitted model
}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec, Necla Kochan,
Gozde Erturk Zararsiz, Selcuk Korkmaz, Gokmen Zararsiz
}
