% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{std.test}
\alias{std.test}
\title{Standardization according to the training model parameters.}
\usage{
std.test(newdata, model)
}
\arguments{
\item{newdata}{a \code{numeric} data frame of biomarkers}

\item{model}{a \code{list} of parameters from the output of linComb,
nonlinComb, mlComb or mathComb functions.}
}
\value{
A \code{numeric} dataframe of standardized biomarkers
}
\description{
The \code{std.test} Standardization parameters will be taken
from the fitted training model and applied to the new data set.
}
\author{
Serra Ilayda Yerlitas, Serra Bersan Gengec, Necla Kochan,
Gozde Erturk Zararsiz, Selcuk Korkmaz, Gokmen Zararsiz
}
