% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{std.train}
\alias{std.train}
\title{Standardization with respect to the t score.}
\usage{
std.train(data, standardize = NULL)
}
\arguments{
\item{data}{a \code{numeric} dataframe of biomarkers}

\item{standardize}{a a \code{string} cdscfvf}
}
\value{
A \code{numeric} dataframe of standardized biomarkers
}
\description{
The \code{std.train} standardizes using T scores. The range
varies between usually 20 and 80
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- exampleData1[, -1]
markers2 <- std.train(markers, tScore)

}
\author{
Serra Ilayda  Yerlitas, Serra Bersan Gengec
}
