% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{std.train}
\alias{std.train}
\title{Standardization according to the chosen method.}
\usage{
std.train(data, standardize = NULL)
}
\arguments{
\item{data}{a \code{numeric} data frame of biomarkers}

\item{standardize}{a \code{character} string indicating the name of the
standardization method. The default option is no standardization applied.
Available options are:
\itemize{
\item \code{range}: Standardization to a range between 0 and 1
\item \code{zScore}: Standardization using z scores with mean = 0
and standard deviation = 1
\item \code{tScore}: Standardization using T scores. The range varies between
usually 20 and 80
\item \code{mean}: Standardization with sample mean = 1
\item \code{deviance}: Standardization with sample standard deviation = 1
}}
}
\value{
A \code{numeric} data.frame of standardized biomarkers
}
\description{
The \code{std.train} Standardization (range, zScore etc.) can be
estimated from the training data and applied to any data set with the same
variables.
}
\examples{
#call data
data(exampleData1)

#define the function parameters
markers <- exampleData1[, -1]
markers2 <- std.train(markers, "deviance")

}
\author{
Serra Ilayda  Yerlitas, Serra Bersan Gengec
}
